SnowflakeMCPAPI/requirements.txt
strands-agents
strands-agents-tools
uv
boto3
bedrock-agentcore
bedrock-agentcore-starter-toolkit
python-dotenv
snowflake-connector-python